% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorR.R
\name{Clusters}
\alias{Clusters}
\title{Preparation for clustering}
\usage{
Clusters(
  data,
  loadings = NULL,
  k0 = 5,
  j0 = ceiling(dim(data)[2]/4),
  r = NULL,
  omega = NULL
)
}
\arguments{
\item{data}{a data frame or matrix}

\item{loadings}{a list containing \code{Loadings_SpecificFactor}. It should be like \code{list(Loadings_SpecificFactor=...)}. If \code{NULL}, the function \code{Loadings_Factor} will be used to calculate the loadings of the factors.}

\item{k0}{a pre-specified integer used in calculating the numbers of factors. It works only when loadings and r is \code{NULL}.}

\item{j0}{a pre-specified integer used in calculating the numbers of factors. It works only when loadings and r is \code{NULL}.}

\item{r}{the numbers of factors. If \code{NULL}, the function \code{FactorNum} will be used to calculate the numbers of the factors.}

\item{omega}{the threshold deciding the components not belonging to any clusters. If \code{NULL}, it will be set to a proper number according to the data and loadings.}
}
\value{
a list consists of \code{NoCluster_Index}, \code{d} and \code{R}. \code{NoCluster_Index} is the indices of the components not belonging to any clusters. \code{d} is the number of clusters. \code{R} is the correlation matrix based on the loading of specific factors, which can be used in k-means clustering.
}
\description{
This function calculates the the components not belonging to any clusters, the number of clusters and the correlation matrix based on the loading of specific factors, which can be used in k-means clustering. See more in \url{https://doi.org/10.1080/01621459.2023.2183132}.
}
\examples{
\dontrun{
data(CSI300_2)
Clusters(CSI300_2)$d
}
}
