% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorR.R
\name{CSI300_1}
\alias{CSI300_1}
\title{Daily closing prices of CSI300 components}
\description{
This dataset comprises the daily closing prices of selected constituent stocks of the CSI 300 Index from January 1, 2018 to December 31, 2023, with missing values subjected to interpolation. The column names correspond to the stock codes.
}
\examples{
\dontrun{
data(CSI300_1)
head(CSI300_1[,1:15])
}
}
