% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorR.R
\name{CSI300_2}
\alias{CSI300_2}
\title{Logarithmic return of CSI300 components}
\description{
This dataset comprises the logarithmic return of the daily closing prices of selected constituent stocks of the CSI 300 Index from January 1, 2018 to December 31, 2023, which is derived from CSI300_1.
}
\examples{
\dontrun{
data(CSI300_2)
head(CSI300_2[,1:15])
}
}
