% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FactorNum_cpp}
\alias{FactorNum_cpp}
\title{Calculating the numbers of factors with Rcpp}
\usage{
FactorNum_cpp(data, k0 = 5L, j0 = -1L)
}
\arguments{
\item{data}{a data frame or matrix}

\item{k0}{a pre-specified integer used in calculating covariance matrix. The exact value won't affect the result much and it could be proper to choose it as 5(the default value) in practice.}

\item{j0}{a pre-specified integer representing the maximum of the number of factors. It must be less than the number of variables. The default value is one forth of the number of variables(rounded down to the nearest integer).}
}
\value{
a list consists of \code{r0}, \code{r}, \code{R} and \code{tao}. \code{r0} is the number of common factors. \code{r} is the number of specific factors. \code{R} is all the ratios of eigenvalue sums. \code{tao} is the indices of local maximums of R, which is sorted into descending order by the corresponding value of R.
}
\description{
Rcpp version of \code{FactorNum}. See more in \code{FactorNum}.
}
\examples{
\dontrun{
data(CSI300_2)
nums <- FactorNum_cpp(CSI300_2)   
cat(c(nums$r0, nums$r))
}
}
