% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorR.R
\name{Loadings_Factor}
\alias{Loadings_Factor}
\title{Calculating the loading matrices of factors}
\usage{
Loadings_Factor(data, k0 = 5, j0 = ceiling(dim(data)[2]/4), r = NULL)
}
\arguments{
\item{data}{a data frame or matrix}

\item{k0}{a pre-specified integer used in calculating the numbers of factors. It works only when r is \code{NULL}.}

\item{j0}{a pre-specified integer used in calculating the numbers of factors. It works only when r is \code{NULL}.}

\item{r}{the numbers of factors. The first value is the numbers of common factors and the second value is the numbers of specific factors. When r is \code{NULL}, the function \code{FactorNum} will be used to calculate the numbers of the factors.}
}
\value{
a list consists of \code{Loadings_CommonFactor} and \code{Loadings_SpecificFactor}, which represent the loadings of common factors and specific factors, respectively.
}
\description{
This function calculates the loading matrices of both common factors and specific factors. See more in \url{https://doi.org/10.1080/01621459.2023.2183132}.
}
\examples{
\dontrun{
data(CSI300_2)
Loadings_Factor(CSI300_2, r=c(1,13))
}
}
