% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Loadings_Factor_cpp}
\alias{Loadings_Factor_cpp}
\title{Calculating the loading matrices of factors with Rcpp}
\usage{
Loadings_Factor_cpp(data, k0 = 5L, j0 = -1L, r = NULL)
}
\arguments{
\item{data}{a data frame or matrix}

\item{k0}{a pre-specified integer used in calculating the numbers of factors. It works only when r is \code{NULL}.}

\item{j0}{a pre-specified integer used in calculating the numbers of factors. It works only when r is \code{NULL}.}

\item{r}{the numbers of factors. The first value is the numbers of common factors and the second value is the numbers of specific factors. When r is \code{NULL}, the function \code{FactorNum} will be used to calculate the numbers of the factors.}
}
\value{
a list consists of \code{Loadings_CommonFactor} and \code{Loadings_SpecificFactor}, which represent the loadings of common factors and specific factors, respectively.
}
\description{
Rcpp version of \code{Loadings_Factor}. See more in \code{Loadings_Factor}.
}
\examples{
\dontrun{
data(CSI300_2)
Loadings_Factor_cpp(CSI300_2)   
}
}
